backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==5.0.2
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
drf-yasg==1.21.7
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
model-bakery==1.17.0
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytest==8.1.1
pytest-django==4.8.0
pytz==2024.1
PyYAML==6.0.1
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
uritemplate==4.1.1
zope.interface==6.3
